implementation of direct queue writer pass
Looks like `ConcurrentWriterCpgPass` is pushing the result DiffGraph in
to writer queue from main thread which is waiting for one by one future
to finish and then push the results to writer queue. This way if one of
task becomes bigger to finish other results will not be written to CPG
unless the bigger one gets finished.

`LinkedBlockingQueue` is concurrent in nature, hence we can push the
DiffGraph to the queue from individual threads/Futures once they are
done with the processing.